Render Slideshow to a Video using FFMPEG.
And ffmpeg process is spawned after images have been loaded with "-i -" so that input come from stdin.

Each frame is rendered into a Framebuffer Texture.

Texture is copied into an Image and that image pixels are sent to ffmpeg as (rgba) via stdin

At the end, stdin is closed so ffmpeg can finish the video properly and the process closes itself.